Add MobileNet training scripts
This set of scripts has two major changes relative to previous work. It uses
MobileNet, a lightweight network intended for use on mobile phones that
has comparable performance but runs much faster and has fewer parameters
than networks such as AlexNet. Also, it uses the new Datasets API for
Tensorflow for reading in data. These scripts train on MNIST data and
will be modified to train on CTA data next.

Note that validation is handled in an unusual way. On each minibatch
update, both a training and validation batch are evaluated (but only the
training batch is used for backprop). This contrasts with the usual
method of occasionally evaluating on the entire validation set, which I
haven't figured out the best way to do yet.