Move image size constants to train code
The width, length, and depth of the images are now read in from the
data, as is the number of telescopes. Previously, they were redundantly
defined in the model code as well, which could lead to inconsistencies.
Since these values are needed to constuct the graph, they are passed to
the model by the training script.