Fix bugs so model trains
Code now runs. Several bugs were fixed including multiplying by the
trig_list correctly and stacking the telescopes along the correct axis
when combining them into the full array.

Because the format of the visualizations relies on single telescope data
and this code is not strictly necessary for training, everything
relating to visualizations and embeddings has been commented out (but
could easily be restored later).