Switch convolutional block to MobileNet
The single telescope convolutional block now uses MobileNet rather than
AlexNet (which remains in the codebase as an option). Note that the
MobileNet block, which is a smaller version of the full original network,
takes up more space than required by the AlexNet block. The version used
here is optimized to permit batches of 64 with 15 telescopes each.

This makes sense because the advantage of switching to MobileNet is to
eliminate the need for large fully connected layers, which haven't been
removed yet. Also, because of the 120x120 image size, it isn't possible
to reduce the size smaller than 15x15 with only stride-2 convolutions,
which results in a limit on how few parameters can be used.

If 128x128 images can be used, the MobileNet 75% version could be a good
version of the network to base our convolutional blocks on.