Combine telescope data into feature map
The final function needed to concatenate single telescope data as
feature maps instead of flattened vectors has been added. Auxiliary
input is included as channels in which every node takes the same value.

The mobilenet_block function was messed up in a previous merge, so it
has been restored to how it was before.

The trig_list is now cast to tf.float32 at the time of input handling by
the model function, as opposed to when the data is read in the training
script.