Pass is_training to cnn_block
The variable indicating whether the model was in training mode wasn't
being passed to the cnn_block, so this section of the model was
presumably being trained on the validation examples. I didn't notice any
differences resulting from this correction, but it could have had subtle
effects that were being masked by other issues.