Compute summaries in same session run as training
This corrects a significant bug in which training batches alternated
between used for training and for calculating metrics. Manual inspection
of batches (with training shuffling turned off) and comparison to the
dataset revealed that only every other training batch was being fed in for
training. Validation batches were not affected. This is because each
call to sess.run advances the training iterator, so that by running
twice on each iteration every other batch was skipped.