Clarify and encapsulate data loading functions
The functions for loading the data have been rewritten to improve the
style and place all logic relating to the specific HDF5 format into
separate functions. This will enable moving these functions into a
separate file if desired. Three functions are required for a given
format to load the data (must take a mode argument specificing train or
eval mode), to load auxiliary data (called once to provide the same
input for every event), and to load metadata (can be used for the other
data loading functions or elsewhere in the program as needed). Metadata
in this context refers to properties of the data such as image width and
height, number of channels, and number of telescopes.