Move metric calculations to train script
The model function now returns only loss and logits. Other metrics, such
as accuracy, AUC, etc. are calculation by train script. The loss
function is needed as part of the model to train it. Other metrics are
employed at the discretion of the user depending on their needs and
should be calculated at a higher level.