Encapsulate model dependencies when parsing config
All logic relating to parameters set when using different models has
been moved to a single location, the configuration file and its parsing.
The biggest change this requires is providing the telescope positions
for every event in the batch when loading data. It is now an option to
sort the telescopes (for the CNN RNN model), but the positions will now
be returned for every event even if that's not the being done and the
positions are all the same. While potentially less efficient, this has
the side effect of simplifying telescope combination, as the positions
don't need to be copied over the batch later.

Options related to the CNN RNN model, ClipGradientNorm and
SortTelescopesByTrigger, have been added to the config file.