Remove dependence of input_fn on data format
The input function is now general and does not rely on the specific
method of mapping (filename, index) pairs to images. The dataset
generation is now general, with the map_func defined by the specific
data format. Reading the list of data files has also been pulled out of
the HDF5 format section.