Make input fn independent of model and data format
The input function now takes a dictionary of settings to enable it to
build the dataset and interpret the iterator output. The settings are
either determined as config options or depend on the data format. This
removes the need to explictly reference the model used and lets the
input fn make fewer implicit assumptions about the nature of the data
(specifically, that mapping is required). The function still assumes
that creating the dataset from a generator is the correct approach. If
this assumption is invalidated by future data formats, how to create the
dataset can be added as another setting later.