Change training to repeated train and evaluation
Perform training using repeated epochs of training on the training set
and evaluation on the validation set, without evaluating on the training
set. After repeated testing and debugging we believe that calling
evaluate twice, on different datasets, without intervening training
results in unexpected behavior. There is no advantage to using
ValidationMonitor, the behavior is the same.