Store file handles in external dictionary
Instead of opening and closing the HDF5 files whenever load_data is
called, instead the files are only opened the first time they're used
with the file handles for each filename being stored in a dictionary.

For single telescope classification, this change gives a ~18x speedup.
For array level classification, since the files are read proportionally
fewer times since all images are read from the same file, the speedup is
much smaller but still significant (~15-20%).

Note that in this implementation, the files are opened for data loading
but never closed. This is tolerable because in practice, we always end
training by killing the process, so file closing would never be
explicitly called anyway.