Pass processed data parameters to model function
This is needed to facilitate reshaping the input data in the model
function, since when cropping is enabled it can no longer be assumed
that the image sizes are the same as defined in the original metadata.
The processed parameters are added to the data processing and metadata
dictionaries. This also increases efficiency since they only need to be
calculated once.