Expand telescope positions to have all aux input
A tensor of telescope auxiliary input is now passed to array-level
models, replacing the previous telescope_positions input. The auxiliary
input is dynamically set depending on the user configuration. The
possibilities are telescope positions and shower positions (when
cropping).

Note: this commit restores casting the triggers to float in array level
models which is necessary for using trigger dropout.