Cut events with no triggered MSTS
Approximately 40% of events have no triggered SCT/MSTS. Hardcode a cut
to remove them for array-level training. This is a temporary fix. A
better implementation is to use a cut string to remove them, but
unfortunately pytables selections cannot operate on array-valued
columns. One solution is to save the number of triggered telescopes of
each type as a column in the event table in the next iteration of the
data format.