Fix errors with predict function
Several sections of code were incomplete and have been added. Labels are
no longer passed to the model scripts, as they are only needed for
calculating the loss (no longer done by the model) and are not provided
when running predict. It is hardcoded to not run in parallel when
running predict (i.e. num_parallel_calls=1), as doing so caused the code
to crash. This is most likely due to file access errors when opening the
HDF5 too many times at once.