Automate training over hyperparameter combinations
It is now possible using the script train_configurations to
systematically train over different hyperparameter combinations.
Hyperparameters may be discrete or continuous, with sampling done either
in a grid or randomly.

To allow training from an external script, train.py has been refactored
so that only reading the config file is handled by main(); logging is
now handled by train(). A change involved in this is that the config
object is now written to a file, as opposed to the original file being
copied directly. This allows calling train() using a config object not
obtained using an actual file.