Clean up config file
Descriptions of the required data type, whether the option is required
or optional, the default setting  and list of all valid options added for all
standard config options (not model-dependent ones) where appropriate.

Data type of options changed to null in cases where None is expected in Python.

String arguments are all quoted. This is not strictly needed but
clarifies the expected data type.

Config option added in Image Mapping for padding, as expected by
ImageMapper.

Logging section moved to the top because in practice this section is
changed for every run and is the primary identifier of the run.