Add image shapes to DataLoader metadata
Fixes bug in which no shape information is present if not using a
DataProcessor. Now DataLoader will provide the normal shapes if a
DataProcessor is not present. In the DataProcessor metadata,
'processed_image_shapes' is renamed to 'image_shapes' so models don't
need to care whether a DataProcessor is being used.