Use only tel_id as a telescope identifier
A fix has been added to handle the incorrect tel_ids for some SST1s due
to an overflow bug in the ImageExtractor data format. Therefore, it's no
longer necessary to use tel_type as a telescope identifier, as discussed
in #39. This allows HDF5DataLoader.selected_telescopes to have the same
format as HDF5DataLoader.telescopes, so that has been implemented.