Upgrade train_configurations script
The script train_configurations.py has been extensively rewritten with
the following major changes:
- Add ability to group together config settings when running over
configurations. This means the script can be used to automate training
over a set of specified configurations, not just for a generic
hyperparameter search.
- Combine the changing_configs file into the base configuration file by
using two special Multiple Configurations sections. These sections are
ignored when simply running a model. This setup is simpler since the
user needs to track fewer files, keeps the settings for a set of
multiple runs all in one place, and takes fuller advantage of YAML.
- Simplify run combinations dictionary file. Instead of storing the full
key for each config option, just save a YAML file mapping a user-provided
config nicknames to values. This is better because the purpose of this
output dictionary is just to give a handy human and machine readable
reference for which run corresponded to which settings without having to
examine the config file in each directory separately. The full config
files for reproducibility are already being stored separately by
run_model.py.