Install TensorFlow with pip
The current method of installing TensorFlow doesn't work properly with
GPU and is not recommended by Google. Instead, install TensorFlow using
pip which is possible using a conda environment YAML file.

In addition, the optional dependencies have been added to the
requirements file. The suppression of TensorFlow INFO messages has been
removed so as to more easily check that GPU is being used.