Fix problems with evaluation and prediction
Fixed bug in which min_num_tels was not being applied.

The validation and test generators are now not shuffled in DataLoader,
which makes checking for reproducibility and using prediction results
easier. Similiarly, the order of files when loading data is now enforced
to be constant.

The calculation of AUC in the TensorFlow evalution metrics has been
corrected to use the classifier values instead of predicted classes.