Improve loading time and update arguments
The loading time was dominated by the creation of virtual pixels, which
was implemented using a triple nested loop. Replacing this with set
operations reduced the loading time from ~70s (SCTCam: ~55s) to ~5s
(SCTCam: ~2s).

In addition, the camera_type argument has been changed to camera_types
to clarify that it should be a list and the default types loaded have
been changed from just LSTCam to all camera types if no argument is
provided.