Refactor run_model and configuration to use DL1DH
Replace DataLoader and ImageMapper with DL1DataReader and use Transforms
to replace DataProcessor. The config file has been rewritten to match
the DL1DH API. In addition, the input_fn and associated config arguments
have been cleaned up to fix #41. In fact, because DL1DH now returns the
examples directly, there is no need to use map at all and prefetching
should be solely relied on for parallelizing the data input pipeline. In
addition, this commit incidentally fixes #40 to allow data files to be
listed directly in the config file.