enable the predict mode for MTL models
I added the function 'write_output()', which writes the output in h5 files (depending on the selection in the config file). I also fixed the format for the different modes. The final evaluation and the predictions follow the 'GammaBoard' format, while the evaluation per epoch follow the 'CTLearn_standard' format.