use mean absoulte error as the metric for the regression tasks
When running the regression tasks in Singletask Learning, the 'mae' metric is equal to the validation loss. The loss of the regression tasks are also computed with 'mae'. However, if you run in Multitask Learning mode, the metrics of the regression tasks differs from the validation loss and therefore the evaluation metrics gain more importance.