added ResNet model and attention mechanisms
This commit added the residual blocks in basic.py needed for the ResNet model (default_models/res_net.py). In addition, basic.py holds the stack_fn() to stack the residual blocks.

Three different attention mechanisms are added in basic.py. Those functions can be used in all CTLearn models.

README and config/example_config.yml are updated accordingly.

This commit removed the gamma mask in the calculation in the regression loss, since it wasn't working properly. This will be tackled in a different commit/branch.