make ctlearn more user-friendly
This commit adds several options to the command line in order to keep the config file as general as possible. As a result we can ship generic config files for each model with the installation of CTLearn. Also the user can select the model from the command line without providing an own config file. Developers can still use all the features as before.